- Remove the requirement to collect the demographic information in sign-up. They should be optional. - Before entering the data, in the sign up page, we have to show the GDPR compliance agreement document. - Remove Honor Education from terms and conditions.
