+++ src/components/SignUpProfessionalInfo.tsx
@@ -1,39 +1,18 @@
-import {
-  Autocomplete,
-  Backdrop,
-  Checkbox,
-  CircularProgress,
-  createFilterOptions,
-  FormHelperText,
-  Link,
-  TextField,
-  Typography,
-} from "@mui/material";
+import { Autocomplete, createFilterOptions, TextField } from "@mui/material";
 import { Box } from "@mui/system";
 import { collection, getDocs, getFirestore, query } from "firebase/firestore";
 import { FormikProps } from "formik";
-import { HTMLAttributes, lazy, Suspense, useEffect, useState } from "react";
+import { HTMLAttributes, useEffect, useState } from "react";
 import { Institution, Major, SignUpFormValues } from "src/knowledgeTypes";
 
 import { EDUCATION_VALUES } from "../lib/utils/constants";
 import OptimizedAvatar from "./OptimizedAvatar";
-
-const CookiePolicy = lazy(() => import("./modals/CookiePolicy"));
-const GDPRPolicy = lazy(() => import("./modals/GDPRPolicy"));
-const PrivacyPolicy = lazy(() => import("./modals/PrivacyPolicy"));
-const TermsOfUse = lazy(() => import("./modals/TermsOfUse"));
-const InformedConsent = lazy(() => import("./modals/InformedConsent"));
 
 type SignUpBasicInformationProps = {
   formikProps: FormikProps<SignUpFormValues>;
 };
 
 export const SignUpProfessionalInfo = ({ formikProps }: SignUpBasicInformationProps) => {
-  const [openInformedConsent, setOpenInformedConsent] = useState(false);
-  const [openGDPRPolicy, setOpenGDPRPolicy] = useState(false);
-  const [openTermOfUse, setOpenTermsOfUse] = useState(false);
-  const [openPrivacyPolicy, setOpenPrivacyPolicy] = useState(false);
-  const [openCookiePolicy, setOpenCookiePolicy] = useState(false);
   const [institutions, setInstitutions] = useState<Institution[]>([]);
   //const [allInstitutions, setAllInstitutions] = useState<Institution[]>([]);
   const [majors, setMajors] = useState<Major[]>([]);
@@ -195,190 +174,6 @@ export const SignUpProfessionalInfo = ({ formikProps }: SignUpBasicInformationPr
         fullWidth
         sx={{ mb: "16px" }}
       />
-
-      {/* <Box></Box> */}
-      <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox checked={values.signUpAgreement} onChange={(_, value) => setFieldValue("signUpAgreement", value)} />
-        <Box>
-          <Box>
-            <Typography>
-              I acknowledge and agree that any data generated from my use of 1Cademy may be utilized for research
-              purposes by the investigators at 1Cademy, the University of Michigan School of Information, and Honor
-              Education.
-            </Typography>
-          </Box>
-          {Boolean(errors.signUpAgreement) && Boolean(touched.signUpAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.signUpAgreement && errors.signUpAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box>
-
-      <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox
-          checked={values.GDPRPolicyAgreement}
-          onChange={(_, value) => setFieldValue("GDPRPolicyAgreement", value)}
-        />
-        <Box>
-          <Box>
-            <Typography>I acknowledge and agree to</Typography>
-            <Link
-              onClick={() => {
-                setFieldValue("clickedGDPR", true);
-                setOpenGDPRPolicy(true);
-              }}
-              sx={{ cursor: "pointer", textDecoration: "none" }}
-            >
-              1Cademy's General Data Protection Regulation (GDPR) Policy.
-            </Link>
-          </Box>
-          {Boolean(errors.GDPRPolicyAgreement) && Boolean(touched.GDPRPolicyAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.GDPRPolicyAgreement && errors.GDPRPolicyAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box>
-
-      <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox
-          checked={values.termsOfServiceAgreement}
-          onChange={(_, value) => setFieldValue("termsOfServiceAgreement", value)}
-        />
-        <Box>
-          <Box>
-            <Typography>I acknowledge and agree to</Typography>
-            <Link
-              onClick={() => {
-                setFieldValue("clickedTOS", true);
-                setOpenTermsOfUse(true);
-              }}
-              sx={{ cursor: "pointer", textDecoration: "none" }}
-            >
-              1Cademy's Terms of Service.
-            </Link>
-          </Box>
-          {Boolean(errors.termsOfServiceAgreement) && Boolean(touched.termsOfServiceAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.termsOfServiceAgreement && errors.termsOfServiceAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box>
-
-      <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox
-          checked={values.privacyPolicyAgreement}
-          onChange={(_, value) => setFieldValue("privacyPolicyAgreement", value)}
-        />
-        <Box>
-          <Box>
-            <Typography>I acknowledge and agree to</Typography>
-            <Link
-              onClick={() => {
-                setFieldValue("clickedPP", true);
-                setOpenPrivacyPolicy(true);
-              }}
-              sx={{ cursor: "pointer", textDecoration: "none" }}
-            >
-              1Cademy's Privacy Policy.
-            </Link>
-          </Box>
-          {Boolean(errors.privacyPolicyAgreement) && Boolean(touched.privacyPolicyAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.privacyPolicyAgreement && errors.privacyPolicyAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box>
-
-      <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox checked={values.cookiesAgreement} onChange={(_, value) => setFieldValue("cookiesAgreement", value)} />
-        <Box>
-          <Box>
-            <Typography>I acknowledge and agree to</Typography>
-            <Link
-              onClick={() => {
-                setFieldValue("clickedCP", true);
-                setOpenCookiePolicy(true);
-              }}
-              sx={{ cursor: "pointer", textDecoration: "none" }}
-            >
-              1Cademy's Cookies Policy.
-            </Link>
-          </Box>
-          {Boolean(errors.cookiesAgreement) && Boolean(touched.cookiesAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.cookiesAgreement && errors.cookiesAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box>
-
-      {/* <Box sx={{ mb: "16px", display: "flex", alignItems: "center" }}>
-        <Checkbox checked={values.ageAgreement} onChange={(_, value) => setFieldValue("ageAgreement", value)} />
-        <Box>
-          <Box>
-            <Typography>I confirm that I am 18 years of age or older.</Typography>
-          </Box>
-          {Boolean(errors.ageAgreement) && Boolean(touched.ageAgreement) && (
-            <FormHelperText sx={{ color: theme => theme.palette.error.main }}>
-              {touched.ageAgreement && errors.ageAgreement}
-            </FormHelperText>
-          )}
-        </Box>
-      </Box> */}
-
-      {/* <Box sx={{ mb: "16px" }}>
-        <Typography>By clicking "Sign Up," you acknowledge that you agree to 1Cademy's </Typography>
-        <Link
-          onClick={() => {
-            setFieldValue("clickedTOS", true);
-            setOpenTermsOfUse(true);
-          }}
-          sx={{ cursor: "pointer", textDecoration: "none" }}
-        >
-          Terms of Use
-        </Link>
-        ,{" "}
-        <Link
-          onClick={() => {
-            setFieldValue("clickedPP", true);
-            setOpenPrivacyPolicy(true);
-          }}
-          sx={{ cursor: "pointer", textDecoration: "none" }}
-        >
-          Privacy Policy
-        </Link>{" "}
-        and{" "}
-        <Link
-          onClick={() => {
-            setFieldValue("clickedCP", true);
-            setOpenCookiePolicy(true);
-          }}
-          sx={{ cursor: "pointer", textDecoration: "none" }}
-        >
-          Cookie Policy
-        </Link>
-        .
-      </Box> */}
-
-      <Suspense
-        fallback={
-          <Backdrop sx={{ color: "#fff", zIndex: theme => theme.zIndex.drawer + 1 }} open={true}>
-            <CircularProgress color="inherit" />
-          </Backdrop>
-        }
-      >
-        <>
-          <InformedConsent open={openInformedConsent} handleClose={() => setOpenInformedConsent(false)} />
-          <GDPRPolicy open={openGDPRPolicy} handleClose={() => setOpenGDPRPolicy(false)} />
-          <CookiePolicy open={openCookiePolicy} handleClose={() => setOpenCookiePolicy(false)} />
-          <PrivacyPolicy open={openPrivacyPolicy} handleClose={() => setOpenPrivacyPolicy(false)} />
-          <TermsOfUse open={openTermOfUse} handleClose={() => setOpenTermsOfUse(false)} />
-        </>
-      </Suspense>
     </Box>
   );
 };
